🐛 Fix: some build errors
