+++ next.config.ts
@@ -2,9 +2,13 @@ import type { NextConfig } from 'next';
 
 const nextConfig: NextConfig = {
   /* config options here */
+
   reactStrictMode: true,
   images: {
     domains: ['sprint-fe-project.s3.ap-northeast-2.amazonaws.com'],
+  },
+  typescript: {
+    ignoreBuildErrors: true,
   },
 };
