+++ styles/board.style.ts
