+++ styles/pagination.style.ts
